# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NUKU API Webhooks
  version: 1.0.0
  description: |-
    API for managing webhook integrations with NUKU-API
    This is a sample Stock Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    Some useful links:
    - [The Stock Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Stock Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
    **Versioning Policy**:  
    - Non-breaking changes (new endpoints/fields) are added to the current version.  
    - Breaking changes require a new version (`v1` → `v2`).  
    - [Changelog](#tag/Changelog)  
  termsOfService: http://api.nuku.com/docs/terms/
  contact:
    email: apiteam@pngx-api.christianaugustyn.me
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: https://api.nuku-api.com.pg/api/v1
    description: Version 1 (Deprecated, use v2)
  - url: https://api.nuku-api.com.pg/api/v2
    description: Version 2 (Latest)
...
paths:
...
  /webhook:
    post:
      tags: [Webhooks]
      summary: Register a new webhook
      description: Create a new webhook subscription for receiving events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookInput'
      responses:
        '201':
          description: Webhook created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          description: Invalid input
          
    get:
      tags: [Webhooks]
      summary: List all webhooks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookListResponse'
...
components:
  schemas:
...
    WebhookInput:
      type: object
      required:
        - endpointUrl
        - eventTypes
      properties:
        endpointUrl:
          type: string
          format: uri
          example: "https://yourdomain.com/webhook"
        eventTypes:
          type: array
          items:
            type: string
            enum: 
              - workflowRun.completed
              - workflowRun.started
              - subscribe
          example: ["workflowRun.completed"]
        workflowId:
          type: string
          nullable: true
        description:
          type: string
    WebhookResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        data:
          $ref: '#/components/schemas/Webhook'
    WebhookListResponse:
      type: object
      properties:
        total:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/Webhook'
    Webhook:
      type: object
      properties:
        id:
          type: string
        endpointUrl:
          type: string
          format: uri
        eventTypes:
          type: array
          items:
            type: string
        workflowId:
          type: string
          nullable: true
        headers:
          type: object
          properties:
            x-cs-signature:
              type: string
            x-cs-timestamp:
              type: integer
            x-webhook-token:
              type: string
        secret:
          type: string
        isActive:
          type: boolean
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    
    # V1 Schema (Legacy)
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
    api_key_auth:
      type: apiKey
      name: X-API-Key
      in: header
security:
  - api_key_auth: []